FIX:  cmd_send has wrong varible definitions, leading to loss of values.

+++ armsrc/apps.h
@@ -230,7 +230,7 @@ void check_challenges(bool file_given, byte_t* data);
 
 // cmd.h
 bool cmd_receive(UsbCommand* cmd);
-bool cmd_send(uint32_t cmd, uint32_t arg0, uint32_t arg1, uint32_t arg2, void* data, size_t len);
+bool cmd_send(uint64_t cmd, uint64_t arg0, uint64_t arg1, uint64_t arg2, void* data, size_t len);
 
 // util.h
 void HfSnoop(int , int);
